+++ doc/README.md
@@ -2,7 +2,7 @@
 
 - [App instance configuration](appInstance.md)
 - [App behavior configuration](appConfiguration.md)
-- [App permissions configuration](appPermissions.md)
+- [Profiles configuration](profiles.md)
 - [Azure functions configuration](apiAzureFunctions.md)
 - [Dashboards](dashboards.md)
 - [Scenario Parameters configuration](scenarioParametersConfiguration.md)
