+++ src/services/config/Permissions.js
@@ -1,0 +1,11 @@
+// Copyright (c) Cosmo Tech.
+// Licensed under the MIT license.
+
+/*
+ * Defines permissions on WebApplication actions
+ */
+export const PERMISSIONS = {
+  canCreateScenario: 'can-create-scenario',
+  canDeleteScenario: 'can-delete-scenario',
+  canEditOrLaunchScenario: 'can-edit-launch-scenario',
+};
